+++ packages/kysely/migrate/migrations/2025-06-05T14:30:00_add_nonce_and_state_to_codes.ts
@@ -1,0 +1,20 @@
+import { Kysely } from "kysely";
+import { Database } from "../../src/db";
+
+export async function up(db: Kysely<Database>): Promise<void> {
+  await db.schema
+    .alterTable("codes")
+    .addColumn("nonce", "varchar(1024)")
+    .execute();
+
+  await db.schema
+    .alterTable("codes")
+    .addColumn("state", "varchar(1024)")
+    .execute();
+}
+
+export async function down(db: Kysely<Database>): Promise<void> {
+  await db.schema.alterTable("codes").dropColumn("nonce").execute();
+
+  await db.schema.alterTable("codes").dropColumn("state").execute();
+}